+++ airflow/example_dags/example_external_task_marker_dag.py
@@ -41,7 +41,7 @@ import datetime
 
 from airflow import DAG
 from airflow.operators.dummy import DummyOperator
-from airflow.sensors.external_task_sensor import ExternalTaskMarker, ExternalTaskSensor
+from airflow.sensors.external_task import ExternalTaskMarker, ExternalTaskSensor
 
 start_date = datetime.datetime(2015, 1, 1)
