+++ dev/provider_packages/refactor_provider_packages.py
@@ -147,7 +147,10 @@ class RefactorBackportPackages:
             ("airflow.operators.python", "airflow.operators.python_operator"),
             ("airflow.sensors.base", "airflow.sensors.base_sensor_operator"),
             ("airflow.sensors.date_time", "airflow.sensors.date_time_sensor"),
+            ("airflow.sensors.external_task", "airflow.sensors.external_task_sensor"),
+            ("airflow.sensors.sql", "airflow.sensors.sql_sensor"),
             ("airflow.sensors.time_delta", "airflow.sensors.time_delta_sensor"),
+            ("airflow.sensors.weekday", "airflow.contrib.sensors.weekday_sensor"),
             ("airflow.utils.session", "airflow.utils.db"),
         ]
         for new, old in changes:
