+++ tests/sensors/test_weekday_sensor.py
@@ -24,7 +24,7 @@ from parameterized import parameterized
 from airflow.exceptions import AirflowSensorTimeout
 from airflow.models import DagBag
 from airflow.models.dag import DAG
-from airflow.sensors.weekday_sensor import DayOfWeekSensor
+from airflow.sensors.weekday import DayOfWeekSensor
 from airflow.utils.timezone import datetime
 from airflow.utils.weekday import WeekDay
 from tests.test_utils import db
